fix: move example extensions code and remove unnecessary dev prefixes
Extensions which are only useful as examples, will not be shown in the frontend.
Some extensions are useful for development, they will still appear with a dev prefix.
