+++ backend/src/localization/i18n/en/texts.json
@@ -53,7 +53,7 @@
       "description": "Open AI LLM integration for reasoning Models"
     },
     "langfuse": {
-      "title": "DEV: Langfuse",
+      "title": "Langfuse",
       "description": "Langfuse integration",
       "baseUrl": "Base url",
       "baseUrlHint": "Base url of langfuse (if self-hosted)",
@@ -65,7 +65,7 @@
       "description": "Mistral LLM integration"
     },
     "ollama": {
-      "title": "DEV: Ollama",
+      "title": "Ollama",
       "description": "Ollama LLM integration"
     },
     "openai": {
@@ -74,7 +74,7 @@
       "endpointHint": "Base url of open ai endpoint"
     },
     "openaiCompatible": {
-      "title": "DEV: OpenAI compatible",
+      "title": "OpenAI compatible",
       "description": "Open AI compatible LLM integration"
     },
     "vertexai": {
@@ -93,10 +93,6 @@
       "title": "Summary Prompt",
       "description": "Defines the summary prompt that is used for conversation label.",
       "promptHint": "Use the {content} placeholder for the actual history."
-    },
-    "always42": {
-      "title": "DEV: Add 42",
-      "description": "Always adds 42 to the sum of two numbers"
     },
     "azureSearch": {
       "title": "Azure AI Search",
@@ -123,10 +119,6 @@
       "title": "Duckduckgo Web Search",
       "description": "Performs a web search using Duckduckgo",
       "maxResults": "Number of search results"
-    },
-    "confirm": {
-      "title": "DEV: Confirm",
-      "description": "Asks for confirmation"
     },
     "azureDalle": {
       "title": "Azure Dall-E",
@@ -193,21 +185,9 @@
       "endpointHint": "Endpoint as URL.",
       "headersHint": "One header per line as Key=Value"
     },
-    "context": {
-      "title": "DEV: Contexts",
-      "description": "Shows the current context"
-    },
-    "simpleInput": {
-      "title": "DEV: Simple Input",
-      "description": "Asks for input from the user"
-    },
     "speechToText": {
       "title": "Speech To Text",
       "description": "Allows speech input via microphone icon"
-    },
-    "userArgs": {
-      "title": "DEV: User Args",
-      "description": "Shows the current user args"
     },
     "filesInConversation": {
       "title": "Search Files in Chat",
@@ -262,6 +242,26 @@
       "value": "Value (Template)",
       "valueHint": "Value that is used during tool call",
       "errorToolCall": "An error occurred during the tool call '{tool}'"
+    },
+    "always42": {
+      "title": "DEV: Add 42",
+      "description": "Always adds 42 to the sum of two numbers"
+    },
+    "confirm": {
+      "title": "DEV: Confirm",
+      "description": "Asks for confirmation"
+    },
+    "context": {
+      "title": "DEV: Contexts",
+      "description": "Shows the current context"
+    },
+    "userArgs": {
+      "title": "DEV: User Args",
+      "description": "Shows the current user args"
+    },
+    "simpleInput": {
+      "title": "DEV: Simple Input",
+      "description": "Asks for input from the user"
     }
   },
   "chat": {
